backend/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
Pillow==10.2.0
transformers==4.37.2
torch==2.2.0
accelerate==0.26.1
pdf2image==1.16.3
pytesseract==0.3.10
python-dotenv==1.0.0
supabase==2.3.4
httpx==0.26.0
pydantic==2.5.3
